Revert "add test framework"
