+++ index.js
@@ -11,9 +11,8 @@
  * `VIDEO_HOST` - The main application URL, used for a convenience link of the GET /api/room route
  */
 var Hapi = require('hapi'),
-    home = require('./lib/api/home'),
-    api = require('./lib/api/room'),
-    signaling = require('./lib/signaling');
+    api = require('./lib/api'),
+    signalingServer = require('./lib/signaling-server');
 
 var port = parseInt(process.env.HTTP_SERVER) || 8080,
     server = new Hapi.Server({
@@ -29,10 +28,16 @@ server.connection({
     port: port
 });
 
-home(server);
+io = signalingServer(server);
+api(server, io);
 
-io = signaling(server);
-api(server.listener, io);
+server.route({
+    method: 'GET',
+    path: '/',
+    handler: function (request, reply) {
+        reply('hello there!');
+    }
+});
 
 server.start(function () {
     console.log('Connection server started on port ' + port);
