--- lib/api/home.js
@@ -1,9 +1,0 @@
-module.exports = function (server) {
-    server.route({
-        method: 'GET',
-        path: '/',
-        handler: function (request, reply) {
-            reply('hello there!');
-        }
-    });
-};
