+++ lib/signaling-server.js
@@ -85,8 +85,8 @@ var handleClient = function (io) {
     };
 };
 
-module.exports = function (http) {
-    var io = socket.listen(http);
+module.exports = function (server) {
+    var io = socket.listen(server.listener);
     io.set('log level', parseInt(process.env.LOGLEVEL) || 1);
     io.sockets.on('connection', handleClient(io));
     return io;
