+++ package.json
@@ -6,7 +6,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index",
-    "test": "./node_modules/.bin/istanbul cover _mocha"
+    "test": "echo \"No test specified\""
   },
   "repository": {
     "type": "git",
@@ -23,16 +23,5 @@
     "lodash": "3.10.1",
     "moniker": "0.1.2",
     "socket.io": "0.9.16"
-  },
-  "devDependencies": {
-    "bluebird": "3.0.5",
-    "chai": "3.4.1",
-    "inject-then": "2.0.3",
-    "istanbul": "0.4.0",
-    "mocha": "2.3.4",
-    "rewire": "2.5.0",
-    "sinon": "1.17.2",
-    "sinon-chai": "2.8.0",
-    "socket.io-client": "0.9.16"
   }
 }
